+++ src/Gyselroth/Helper/HelperMongoDB.php
@@ -22,8 +22,7 @@ class HelperMongoDB
      */
     public static function fileExists(string $fileName): bool
     {
-        return (bool)(new MongoDb_Contents())
-            ->getFile($fileName);
+        return (bool)(new MongoDb_Contents())->getFile($fileName);
     }
 
     /**
@@ -45,8 +44,7 @@ class HelperMongoDB
             $fileName = \str_replace(' ', '_', $fileName);
         }
 
-        return (bool)(new MongoDb_Contents())
-            ->storeFile($fileName, \file_get_contents($tmpFileName));
+        return (bool)(new MongoDb_Contents())->storeFile($fileName, \file_get_contents($tmpFileName));
     }
 
     /**
@@ -61,8 +59,7 @@ class HelperMongoDB
             $fileName = \str_replace(' ', '_', $fileName);
         }
 
-        return (bool)(new MongoDb_Contents())
-            ->removeFile($fileName);
+        return (bool)(new MongoDb_Contents())->removeFile($fileName);
     }
 
     /**
@@ -74,8 +71,7 @@ class HelperMongoDB
      */
     public static function doesFileExist(string $fileName): bool
     {
-        return (bool)(new MongoDb_Contents())
-            ->getFile($fileName);
+        return (bool)(new MongoDb_Contents())->getFile($fileName);
     }
 
     /**
@@ -88,7 +84,6 @@ class HelperMongoDB
      */
     public static function createFile(string $fileName, string $tmpFileName): bool
     {
-        return (bool)(new MongoDb_Contents())
-            ->storeFile($fileName, \file_get_contents($tmpFileName));
+        return (bool)(new MongoDb_Contents())->storeFile($fileName, \file_get_contents($tmpFileName));
     }
 }